akencoder: separate AudioBuffer from producer.cpp
